ref:Move `CustomHTTPHandler` code out of `config.py`
Move `CustomHTTPHandler` code out of `config.py` into `custom_http_handler.py` to better
segregate features.

Closes #69.
